refactor(expert-hub): extract viz logic to components/expert_hub_page/viz.js
