+++ foundation_cms/static/scss/pages/expert_hub_page.scss
@@ -117,7 +117,7 @@ body.template-expert-hub-page {
       border-radius: rem-calc(8);
       padding: rem-calc(16) rem-calc(20);
       max-width: rem-calc(280);
-      filter: drop-shadow(0 rem-calc(4) rem-calc(16) rgba(0, 0, 0, 0.12));
+      filter: drop-shadow(0 rem-calc(4) rem-calc(16) rgb(0 0 0 / 12%));
       pointer-events: none;
 
       &__quote {
